Added Settings Menu & Additional Logic
Did a load of work with the settings and being able to specify if CSV /
HTML output files are generated or not. Won't save when you leave the
script yet but it's a work in progress. Does the job for now! Need to
specify timeouts next. Maybe attempt to find the page 2/3 times and
lower timeout time? Maybe check main page first to see if it is "live"
if so then it means user cookie (Pass) is wrong.